API 수정
- 로그인 API Methods 변경 (POST)
- 뉴스피드 & 검색 포스트 반환 값 수정
- 검색 포스트 반환 Oid 반환
- 공지사항 작성자(SIGNUS) 반환
- 검색 정렬 기능 추가
- 조회 API 버그 수정
- TEST MODULE 순서 수정

+++ SIGNUS/app/api/signus_v1/management.py
@@ -4,7 +4,7 @@ SIGNUS V1 management API
 from flask import g, request
 from app.api import input_check
 from app.api.signus_v1 import signus_v1 as api
-from app.api.decorators import timer, login_required, admin_required
+from app.api.decorators import timer, login_required
 from app.controllers.management import (get_notice,
                                      insert_notice,
                                      update_notice,
@@ -16,6 +16,7 @@ from app.controllers.management import (get_notice,
 @timer
 def signus_v1_get_notice(notice_oid=None):
     ''' 공지 반환 (인자가 안들어오면 전체 반환) '''
+
     return {
         "msg": "success",
         "result": get_notice(g.mongo_cur, notice_oid)
@@ -24,28 +25,39 @@ def signus_v1_get_notice(notice_oid=None):
 
 @api.route("/notice", methods=["PUT"])
 @timer
-@admin_required
+@login_required
 def signus_v1_put_notice():
     ''' 공지 추가 '''
+    
+    if not g.user['admin']:
+        return {"msg": "Bad Access Token"}, 401
+    
     data = request.get_json()
     input_check(data, 'title', str)
     input_check(data, 'post', str)
+
     return {
         "msg": "success",
         "result": insert_notice(g.mongo_cur,
                                 data['title'],
-                                data['post'])
+                                data['post'],
+                                g.user)
     }
 
 
 @api.route("/notice/<string:notice_oid>", methods=["PATCH"])
 @timer
-@admin_required
+@login_required
 def signus_v1_patch_notice(notice_oid=None):
     ''' 공지 수정 '''
+    
+    if not g.user['admin']:
+        return {"msg": "Bad Access Token"}, 401
+    
     data = request.get_json()
     input_check(data, 'title', str)
     input_check(data, 'post', str)
+    
     return {
         "msg": "success",
         "result": update_notice(g.mongo_cur,
@@ -57,9 +69,13 @@ def signus_v1_patch_notice(notice_oid=None):
 
 @api.route("/notice/<string:notice_oid>", methods=["DELETE"])
 @timer
-@admin_required
+@login_required
 def signus_v1_delete_notice(notice_oid=None):
     ''' 공지 삭제 '''
+
+    if not g.user['admin']:
+        return {"msg": "Bad Access Token"}, 401
+
     return {
         "msg": "success",
         "result": delete_notice(g.mongo_cur,
